requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
Django==3.2
djangorestframework==3.12.4
httpie==2.4.0
idna==2.10
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
Pygments==2.8.1
PySocks==1.7.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-toolbelt==0.9.1
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.4
